fill_objects and format_line dabuter.

+++ Headers/RT_headers/camera.h
@@ -6,7 +6,7 @@
 /*   By: alvmoral <alvmoral@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/20 12:50:09 by ggasset-          #+#    #+#             */
-/*   Updated: 2025/09/23 18:44:28 by alvmoral         ###   ########.fr       */
+/*   Updated: 2025/10/22 11:41:41 by alvmoral         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,8 +15,7 @@
 
 # include "objects.h"
 # include "stdlib.h"
-// # include "../minilibx_funcs.h"
-#include <stdio.h>
+# include <stdio.h>
 
 typedef struct t_light
 {
@@ -54,18 +53,18 @@ t_color	get_sky_color(t_camera camera, size_t pixel_y);
 t_ray	create_ray(t_camera camera, size_t pixel_i);
 // int		render(t_render_data *render_data);
 
-void			print_t_scene(t_scene s);
-char			**format_line(char *line);
-unsigned int	ft_atoui(const char *str);
-t_data			get_scalar(char *value);
-size_t			get_scalar_size_t(char *value);
-t_vec3			get_vector(char *dim);
+void	print_t_scene(t_scene s);
+char	**format_line(char *line);
+size_t	ft_atoui(const char *str);
+t_data	get_scalar(char *value);
+size_t	get_scalar_size_t(char *value);
+t_vec3	get_vector(char *dim);
 
-int				fill_dispatcher(t_scene *s, char *line);
-void			fill_ambient_ligth(t_light *ambient, char **argv);
-void			fill_ligth(t_light *light, char **argv);
-void			fill_camera(t_camera *cam, char **argv);
-void			fill_obj(t_object *obj, char **argv);
-void			fill_obj_list(t_object_list *objs, char **argv);
+int		fill_dispatcher(t_scene *s, char *line);
+void	fill_ambient_ligth(t_light *ambient, char **argv);
+void	fill_ligth(t_light *light, char **argv);
+void	fill_camera(t_camera *cam, char **argv);
+void	fill_obj(t_object *obj, char **argv);
+void	fill_obj_list(t_object_list *objs, char **argv);
 
 #endif
